+++ data/duke.txt
@@ -1,3 +1,3 @@
-D | 0 | work | 22-01-2024 1800
-D | 0 | do WORK | 09-02-2024 2359
-E | 0 | birthday | 22-06-2024 1330 | 23-06-2024 1545
+E | 0 | valentine's day | 14-02-2024 1930 | 14-02-2024 2200
+D | 1 | buy gift | 17-08-2024 1800
+0 | return money
